Update links of node and express icons to fix anchor tags
Clicking other icons takes you to the site of that icon, however this was broken with ndoe and express icons since I added custom png (with white background). When these icons are clicked, the user is instead taken to the icon png, instead of the node or express site. To fix this I hosted the pngs in this repo and changed the links to the hosted pngs.